requirements.txt
# Testing dependencies
pytest>=7.0.0

# Agent dependencies
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pydantic>=2.5.0
